requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
faker==20.1.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
tokenizers==0.15.0
huggingface-hub==0.19.4
safetensors==0.4.0
regex==2023.10.3
pillow==10.1.0
tqdm==4.66.1
filelock==3.13.1
fsspec==2023.10.0
packaging==23.2
pyyaml==6.0.1
sympy==1.12
networkx==3.2.1
jinja2==3.1.2
mpmath==1.3.0